net.benbenng.wwwapi/requirements.txt
flask
python-sql
mysql-connector
redis